[skip actions] lab/lab01/ch01_t13_value_error.py

+++ lab/lab01/ch01_t13_value_error.py
@@ -1,5 +1,5 @@
 float_1 = 0.25
 float_2 = 40.0
-float product=float_1*float_2
+int product=float_1*float_2
